pyproject.toml
[project]
name = "personal-ai-os"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "pytz>=2025.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.2",
    "telethon>=1.42.0",
]
